masp_proofs
Module constants

Various constants used for the MASP proofs.

Structs

NOTE_COMMITMENT_RANDOMNESS_GENERATOR

NULLIFIER_POSITION_GENERATOR

PEDERSEN_CIRCUIT_GENERATORS

The pre-computed window tables [-4, 3, 2, 1, 1, 2, 3, 4] of different magnitudes of the Pedersen hash segment generators.

PROOF_GENERATION_KEY_GENERATOR

SPENDING_KEY_GENERATOR

VALUE_COMMITMENT_RANDOMNESS_GENERATOR

Functions

generate_circuit_generator

Creates the 3-bit window table [0, 1, ..., 8] for different magnitudes of a fixed generator.

Type Aliases

FixedGenerator

Reference to a circuit version of a generator for fixed-base salar multiplication.

FixedGeneratorOwned

Circuit version of a generator for fixed-base salar multiplication.