Crate masp_proofs

MASP circuits and proofs.

masp_proofs contains the zk-SNARK circuits used by MASP based on Zcash Sapling, and the APIs for creating and verifying proofs.

Re-exports

pub use bellman;

pub use bls12_381;

pub use group;

pub use jubjub;

Modules

circuit

Implementations of the MASP circuits and personalized Pedersen hash

constants

Various constants used for the MASP proofs.

hashreader

Abstraction over a reader which hashes the data being read.

proverlocal-prover or bundled-prover

Abstractions over the proving system and parameters for ease of use.

sapling

Helpers for creating MASP Sapling proofs.

Structs

MASPParameterPathsdownload-params

The paths to the Sapling parameter files.

MASPParameters

MASP Sapling groth16 circuit parameters.

Constants

MASP_CONVERT_BYTES

MASP_CONVERT_HASH

MASP_CONVERT_NAME

The MASP convert parameters file name.

MASP_OUTPUT_BYTES

MASP_OUTPUT_HASH

MASP_OUTPUT_NAME

The MASP output parameters file name.

MASP_SPEND_BYTES

MASP_SPEND_HASH

MASP_SPEND_NAME

The MASP spend parameters file name.

Functions

default_params_folderdirectories

Returns the default folder that the MASP proving parameters are located in.

download_masp_parametersdownload-params

Download the MASP parameters if needed, and store them in the default location. Always checks the sizes and hashes of the files, even if they didn’t need to be downloaded.

load_parameters

Load the specified parameters, checking the sizes and hashes of the files.

parse_parameters

Parse Bls12 keys from bytes as serialized by Parameters::write.